Add dqlite compaction test
